Local spot GUI redesign Bools are written to pts as int Skip disabled spots when procesing Fixed missing processing runs: en/disabling spots, changing spot curve

+++ Sources/imagespot/ptImageSpot.cpp
@@ -89,7 +89,7 @@ void ptImageSpot::setPos(uint Ax, uint Ay) {
 //==============================================================================
 
 void ptImageSpot::WriteToFile(QSettings *APtsFile) {
-  APtsFile->setValue("IsEnabled", FIsEnabled);
+  APtsFile->setValue("IsEnabled", (int)FIsEnabled);
   APtsFile->setValue("Name", FName);
   APtsFile->setValue("Radius", FRadius);
   APtsFile->setValue("SpotPosX", FPos.x());
